refactor(storage): share advisory file locking
